@@ -2,8 +2,9 @@
   "author": "Jaco Greeff <jacogr@gmail.com>",
   "bugs": "https://github.com/polkadot-js/apps/issues",
   "description": "An Apps portal into the Polkadot network",
+  "engine-strict": true,
   "engines": {
-    "node": ">=14"
+    "node": ">=18.14"
   },
   "homepage": "https://github.com/polkadot-js/apps#readme",
   "license": "Apache-2.0",
@@ -15,9 +16,9 @@
     "url": "https://github.com/polkadot-js/apps.git"
   },
   "sideEffects": false,
-  "version": "0.124.2-174-x",
+  "version": "0.124.2-192-x",
   "versions": {
-    "git": "0.124.2-174-x",
+    "git": "0.124.2-192-x",
     "npm": "0.124.1"
   },
   "workspaces": [
@@ -69,10 +70,10 @@
   "devDependencies": {
     "@crustio/crust-pin": "^1.0.0",
     "@pinata/sdk": "^2.1.0",
-    "@polkadot/dev": "^0.72.15",
+    "@polkadot/dev": "^0.72.25",
     "@types/chart.js": "^2.9.37",
     "@types/file-saver": "^2.0.5",
-    "@types/react-beautiful-dnd": "^13.1.3",
+    "@types/react-beautiful-dnd": "^13.1.4",
     "@types/react-copy-to-clipboard": "^5.0.4",
     "@types/react-dom": "^18.0.11",
     "@types/react-router-dom": "^5.3.3",
@@ -82,10 +83,10 @@
     "concurrently": "^7.6.0",
     "devtron": "^1.4.0",
     "dnslink-cloudflare": "^3.0.0",
-    "electron": "^23.1.4",
+    "electron": "^23.2.0",
     "electron-builder": "23.6.0",
     "electron-builder-notarize": "^1.5.1",
-    "i18next-scanner": "^4.1.1",
+    "i18next-scanner": "^4.2.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-is": "^18.2.0",
@@ -93,34 +94,34 @@
     "styled-components": "^5.3.9"
   },
   "resolutions": {
-    "@polkadot/api": "^10.1.3",
-    "@polkadot/api-augment": "^10.1.3",
-    "@polkadot/api-base": "^10.1.3",
-    "@polkadot/api-contract": "^10.1.3",
-    "@polkadot/api-derive": "^10.1.3",
-    "@polkadot/hw-ledger": "^11.0.2",
-    "@polkadot/keyring": "^11.0.2",
-    "@polkadot/networks": "^11.0.2",
-    "@polkadot/phishing": "^0.20.2",
-    "@polkadot/rpc-augment": "^10.1.3",
-    "@polkadot/rpc-core": "^10.1.3",
-    "@polkadot/rpc-provider": "^10.1.3",
-    "@polkadot/types": "^10.1.3",
-    "@polkadot/types-augment": "^10.1.3",
-    "@polkadot/types-codec": "^10.1.3",
-    "@polkadot/types-create": "^10.1.3",
-    "@polkadot/types-known": "^10.1.3",
-    "@polkadot/types-support": "^10.1.3",
-    "@polkadot/util": "^11.0.2",
-    "@polkadot/util-crypto": "^11.0.2",
+    "@polkadot/api": "^10.2.1",
+    "@polkadot/api-augment": "^10.2.1",
+    "@polkadot/api-base": "^10.2.1",
+    "@polkadot/api-contract": "^10.2.1",
+    "@polkadot/api-derive": "^10.2.1",
+    "@polkadot/hw-ledger": "^11.1.2",
+    "@polkadot/keyring": "^11.1.2",
+    "@polkadot/networks": "^11.1.2",
+    "@polkadot/phishing": "^0.20.4",
+    "@polkadot/rpc-augment": "^10.2.1",
+    "@polkadot/rpc-core": "^10.2.1",
+    "@polkadot/rpc-provider": "^10.2.1",
+    "@polkadot/types": "^10.2.1",
+    "@polkadot/types-augment": "^10.2.1",
+    "@polkadot/types-codec": "^10.2.1",
+    "@polkadot/types-create": "^10.2.1",
+    "@polkadot/types-known": "^10.2.1",
+    "@polkadot/types-support": "^10.2.1",
+    "@polkadot/util": "^11.1.2",
+    "@polkadot/util-crypto": "^11.1.2",
     "@polkadot/wasm-crypto": "^7.0.3",
-    "@polkadot/x-bigint": "^11.0.2",
-    "@polkadot/x-fetch": "^11.0.2",
-    "@polkadot/x-global": "^11.0.2",
-    "@polkadot/x-randomvalues": "^11.0.2",
-    "@polkadot/x-textdecoder": "^11.0.2",
-    "@polkadot/x-textencoder": "^11.0.2",
-    "@polkadot/x-ws": "^11.0.2",
+    "@polkadot/x-bigint": "^11.1.2",
+    "@polkadot/x-fetch": "^11.1.2",
+    "@polkadot/x-global": "^11.1.2",
+    "@polkadot/x-randomvalues": "^11.1.2",
+    "@polkadot/x-textdecoder": "^11.1.2",
+    "@polkadot/x-textencoder": "^11.1.2",
+    "@polkadot/x-ws": "^11.1.2",
     "styled-components": "^5.3.1",
     "typescript": "^5.0.2"
   }
