@@ -4,7 +4,7 @@
 import type { RouteProps } from '@polkadot/apps-routing/types';
 
 import React from 'react';
-import { Redirect } from 'react-router';
+import { Navigate } from 'react-router';
 
 interface Props extends RouteProps {
   missingApis?: (string | string[])[];
@@ -14,7 +14,7 @@ function NotFound ({ basePath, missingApis = [] }: Props): React.ReactElement {
   console.log(`Redirecting from route "${basePath}" to "/explorer"${missingApis.length ? `, missing the following APIs: ${JSON.stringify(missingApis)}` : ''}`);
 
   return (
-    <Redirect to='/explorer' />
+    <Navigate to='/explorer' />
   );
 }
 
