@@ -4,7 +4,7 @@
 import type { AppProps as Props } from '@polkadot/react-components/types';
 
 import React, { useRef } from 'react';
-import { Route, Switch } from 'react-router';
+import { Route, Routes } from 'react-router';
 
 import { Tabs } from '@polkadot/react-components';
 
@@ -28,11 +28,16 @@ function AddressesApp ({ basePath, onStatusChange }: Props): React.ReactElement<
         basePath={basePath}
         items={itemsRef.current}
       />
-      <Switch>
-        <Route>
-          <Contacts onStatusChange={onStatusChange} />
+      <Routes>
+        <Route path={basePath}>
+          <Route
+            element={
+              <Contacts onStatusChange={onStatusChange} />
+            }
+            index
+          />
         </Route>
-      </Switch>
+      </Routes>
     </main>
   );
 }
