@@ -6,7 +6,7 @@ import type { TabItem } from '@polkadot/react-components/Tabs/types';
 import type { KeyedEvent } from '@polkadot/react-hooks/ctx/types';
 
 import React, { useMemo, useRef } from 'react';
-import { Route, Switch } from 'react-router';
+import { Route, Routes } from 'react-router';
 
 import { Tabs } from '@polkadot/react-components';
 import { useApi, useBlockAuthors, useBlockEvents } from '@polkadot/react-hooks';
@@ -24,19 +24,6 @@ interface Props {
   basePath: string;
   className?: string;
   newEvents?: KeyedEvent[];
-}
-
-function createPathRef (basePath: string): Record<string, string | string[]> {
-  return {
-    api: `${basePath}/api`,
-    forks: `${basePath}/forks`,
-    latency: `${basePath}/latency`,
-    node: `${basePath}/node`,
-    query: [
-      `${basePath}/query/:value`,
-      `${basePath}/query/`
-    ]
-  };
 }
 
 function createItemsRef (t: TFunction): TabItem[] {
@@ -77,7 +64,6 @@ function ExplorerApp ({ basePath, className }: Props): React.ReactElement<Props>
   const { lastHeaders } = useBlockAuthors();
   const { eventCount, events } = useBlockEvents();
   const itemsRef = useRef(createItemsRef(t));
-  const pathRef = useRef(createPathRef(basePath));
 
   const hidden = useMemo<string[]>(
     () => isFunction(api.query.babe?.authorities) ? [] : ['forks'],
@@ -91,20 +77,40 @@ function ExplorerApp ({ basePath, className }: Props): React.ReactElement<Props>
         hidden={hidden}
         items={itemsRef.current}
       />
-      <Switch>
-        <Route path={pathRef.current.api}><Api /></Route>
-        <Route path={pathRef.current.forks}><Forks /></Route>
-        <Route path={pathRef.current.latency}><Latency /></Route>
-        <Route path={pathRef.current.query}><BlockInfo /></Route>
-        <Route path={pathRef.current.node}><NodeInfo /></Route>
-        <Route>
-          <Main
-            eventCount={eventCount}
-            events={events}
-            headers={lastHeaders}
+      <Routes>
+        <Route path={basePath}>
+          <Route
+            element={<Api />}
+            path='api'
+          />
+          <Route
+            element={<Forks />}
+            path='forks'
+          />
+          <Route
+            element={<Latency />}
+            path='latency'
+          />
+          <Route
+            element={<NodeInfo />}
+            path='node'
+          />
+          <Route
+            element={<BlockInfo />}
+            path='query/:value?'
+          />
+          <Route
+            element={
+              <Main
+                eventCount={eventCount}
+                events={events}
+                headers={lastHeaders}
+              />
+            }
+            index
           />
         </Route>
-      </Switch>
+      </Routes>
     </main>
   );
 }
