@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import React, { useRef } from 'react';
-import { Route, Switch } from 'react-router';
+import { Route, Routes } from 'react-router';
 
 import { Tabs } from '@polkadot/react-components';
 
@@ -31,11 +31,16 @@ function GiltApp ({ basePath, className }: Props): React.ReactElement<Props> {
         basePath={basePath}
         items={tabsRef.current}
       />
-      <Switch>
-        <Route>
-          <Overview />
+      <Routes>
+        <Route path={basePath}>
+          <Route
+            element={
+              <Overview />
+            }
+            index
+          />
         </Route>
-      </Switch>
+      </Routes>
     </main>
   );
 }
