@@ -4,7 +4,7 @@
 import '@polkadot/api-augment/substrate';
 
 import React, { useMemo, useRef } from 'react';
-import { Route, Switch } from 'react-router';
+import { Route, Routes } from 'react-router';
 
 import { Tabs } from '@polkadot/react-components';
 import { useAccounts } from '@polkadot/react-hooks';
@@ -52,17 +52,25 @@ function NftApp ({ basePath, className }: Props): React.ReactElement<Props> {
         hidden={hidden}
         items={tabsRef.current}
       />
-      <Switch>
-        <Route path={`${basePath}/my-nfts`}>
-          <AccountItems infos={infos} />
-        </Route>
-        <Route>
-          <Overview
-            ids={ids}
-            infos={infos}
+      <Routes>
+        <Route path={basePath}>
+          <Route
+            element={
+              <AccountItems infos={infos} />
+            }
+            path='my-nfts'
+          />
+          <Route
+            element={
+              <Overview
+                ids={ids}
+                infos={infos}
+              />
+            }
+            index
           />
         </Route>
-      </Switch>
+      </Routes>
     </main>
   );
 }
