@@ -6,7 +6,7 @@ import '@polkadot/api-augment/substrate';
 import type { ParaId } from '@polkadot/types/interfaces';
 
 import React, { useRef } from 'react';
-import { Route, Switch } from 'react-router';
+import { Route, Routes } from 'react-router';
 import { useLocation } from 'react-router-dom';
 
 import { Tabs } from '@polkadot/react-components';
@@ -76,27 +76,38 @@ function ParachainsApp ({ basePath, className }: Props): React.ReactElement<Prop
         basePath={basePath}
         items={items.current}
       />
-      <Switch>
-        <Route path={`${basePath}/auctions`}>
-          <Auctions
-            auctionInfo={auctionInfo}
-            campaigns={campaigns}
-            ownedIds={ownedIds}
-            winningData={winningData}
+      <Routes>
+        <Route path={basePath}>
+          <Route
+            element={
+              <Auctions
+                auctionInfo={auctionInfo}
+                campaigns={campaigns}
+                ownedIds={ownedIds}
+                winningData={winningData}
+              />
+            }
+            path='auctions'
+          />
+          <Route
+            element={
+              <Crowdloan
+                auctionInfo={auctionInfo}
+                campaigns={campaigns}
+                leasePeriod={leasePeriod}
+                ownedIds={ownedIds}
+              />
+            }
+            path='crowdloan'
+          />
+          <Route
+            element={
+              <Proposals proposals={proposals} />
+            }
+            path='proposals'
           />
         </Route>
-        <Route path={`${basePath}/crowdloan`}>
-          <Crowdloan
-            auctionInfo={auctionInfo}
-            campaigns={campaigns}
-            leasePeriod={leasePeriod}
-            ownedIds={ownedIds}
-          />
-        </Route>
-        <Route path={`${basePath}/proposals`}>
-          <Proposals proposals={proposals} />
-        </Route>
-      </Switch>
+      </Routes>
       <Overview
         actionsQueue={actionsQueue}
         className={pathname === basePath ? '' : '--hidden'}
