@@ -4,7 +4,7 @@
 import type { PalletColl, PalletPoll } from './types.js';
 
 import React, { useMemo } from 'react';
-import { Route, Switch } from 'react-router';
+import { Route, Routes } from 'react-router';
 
 import Referenda, { useCounter } from '@polkadot/app-referenda/Referenda';
 import { Tabs } from '@polkadot/react-components';
@@ -46,19 +46,27 @@ function App ({ basePath, className, palletColl, palletPoll }: Props): React.Rea
         basePath={basePath}
         items={tabs}
       />
-      <Switch>
-        <Route path={`${basePath}/referenda`}>
-          <Referenda
-            members={members && members.memberIds}
-            palletReferenda={palletPoll}
-            palletVote={palletColl}
-            ranks={members && members.memberRanks}
+      <Routes>
+        <Route path={basePath}>
+          <Route
+            element={
+              <Referenda
+                members={members && members.memberIds}
+                palletReferenda={palletPoll}
+                palletVote={palletColl}
+                ranks={members && members.memberRanks}
+              />
+            }
+            path='referenda'
+          />
+          <Route
+            element={
+              <Members members={members && members.members} />
+            }
+            index
           />
         </Route>
-        <Route>
-          <Members members={members && members.members} />
-        </Route>
-      </Switch>
+      </Routes>
     </main>
   );
 }
