@@ -4,7 +4,7 @@
 import type { AppProps as Props } from '@polkadot/react-components/types';
 
 import React, { useRef } from 'react';
-import { Route, Switch } from 'react-router';
+import { Route, Routes } from 'react-router';
 
 import { Tabs } from '@polkadot/react-components';
 
@@ -33,10 +33,22 @@ function SigningApp ({ basePath }: Props): React.ReactElement<Props> {
         basePath={basePath}
         items={itemsRef.current}
       />
-      <Switch>
-        <Route path={`${basePath}/verify`}><Verify /></Route>
-        <Route><Sign /></Route>
-      </Switch>
+      <Routes>
+        <Route path={basePath}>
+          <Route
+            element={
+              <Verify />
+            }
+            path='verify'
+          />
+          <Route
+            element={
+              <Sign />
+            }
+            index
+          />
+        </Route>
+      </Routes>
     </main>
   );
 }
