@@ -5,7 +5,7 @@ import type { DeriveSociety, DeriveSocietyMember } from '@polkadot/api-derive/ty
 import type { MapMember } from './types.js';
 
 import React, { useMemo } from 'react';
-import { Route, Switch } from 'react-router';
+import { Route, Routes } from 'react-router';
 
 import { Tabs } from '@polkadot/react-components';
 import { useApi, useCall } from '@polkadot/react-hooks';
@@ -118,28 +118,39 @@ function SocietyApp ({ basePath, className }: Props): React.ReactElement<Props> 
         basePath={basePath}
         items={items}
       />
-      <Switch>
-        <Route path={`${basePath}/candidates`}>
-          <Candidates
-            allMembers={allMembers}
-            candidates={candidates}
-            isMember={isMember}
-            ownMembers={ownMembers}
+      <Routes>
+        <Route path={basePath}>
+          <Route
+            element={
+              <Candidates
+                allMembers={allMembers}
+                candidates={candidates}
+                isMember={isMember}
+                ownMembers={ownMembers}
+              />
+            }
+            path='candidates'
+          />
+          <Route
+            element={
+              <Suspended />
+            }
+            path='suspended'
+          />
+          <Route
+            element={
+              <Overview
+                info={info}
+                isMember={isMember}
+                mapMembers={mapMembers}
+                ownMembers={ownMembers}
+                payoutTotal={payoutTotal}
+              />
+            }
+            index
           />
         </Route>
-        <Route path={`${basePath}/suspended`}>
-          <Suspended />
-        </Route>
-        <Route>
-          <Overview
-            info={info}
-            isMember={isMember}
-            mapMembers={mapMembers}
-            ownMembers={ownMembers}
-            payoutTotal={payoutTotal}
-          />
-        </Route>
-      </Switch>
+      </Routes>
     </main>
   );
 }
