@@ -5,7 +5,7 @@ import type { CollectiveType } from '@polkadot/react-hooks/types';
 import type { Hash } from '@polkadot/types/interfaces';
 
 import React, { useMemo } from 'react';
-import { Route, Switch } from 'react-router';
+import { Route, Routes } from 'react-router';
 
 import { Tabs } from '@polkadot/react-components';
 import { useApi, useCall, useCollectiveMembers } from '@polkadot/react-hooks';
@@ -53,26 +53,34 @@ function TechCommApp ({ basePath, className, type }: Props): React.ReactElement<
         }
         items={items}
       />
-      <Switch>
-        <Route path={`${basePath}/proposals`}>
-          <Proposals
-            isMember={isMember}
-            members={members}
-            prime={prime}
-            proposalHashes={proposalHashes}
-            type={type}
-          />
-        </Route>
+      <Routes>
         <Route path={basePath}>
-          <Overview
-            isMember={isMember}
-            members={members}
-            prime={prime}
-            proposalHashes={proposalHashes}
-            type={type}
+          <Route
+            element={
+              <Proposals
+                isMember={isMember}
+                members={members}
+                prime={prime}
+                proposalHashes={proposalHashes}
+                type={type}
+              />
+            }
+            path='proposals'
+          />
+          <Route
+            element={
+              <Overview
+                isMember={isMember}
+                members={members}
+                prime={prime}
+                proposalHashes={proposalHashes}
+                type={type}
+              />
+            }
+            index
           />
         </Route>
-      </Switch>
+      </Routes>
     </main>
   );
 }
