@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import React, { useMemo } from 'react';
-import { Route, Switch } from 'react-router';
+import { Route, Routes } from 'react-router';
 
 import { Tabs } from '@polkadot/react-components';
 import { useApi, useCollectiveMembers } from '@polkadot/react-hooks';
@@ -51,21 +51,29 @@ function TreasuryApp ({ basePath }: Props): React.ReactElement<Props> {
         basePath={basePath}
         items={items}
       />
-      <Switch>
-        <Route path={`${basePath}/tips`}>
-          <Tips
-            hashes={tipHashes}
-            isMember={isMember}
-            members={members}
+      <Routes>
+        <Route path={basePath}>
+          <Route
+            element={
+              <Tips
+                hashes={tipHashes}
+                isMember={isMember}
+                members={members}
+              />
+            }
+            path='tips'
+          />
+          <Route
+            element={
+              <Overview
+                isMember={isMember}
+                members={members}
+              />
+            }
+            index
           />
         </Route>
-        <Route>
-          <Overview
-            isMember={isMember}
-            members={members}
-          />
-        </Route>
-      </Switch>
+      </Routes>
     </main>
   );
 }
