@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import React, { useRef } from 'react';
-import { Route, Switch } from 'react-router';
+import { Route, Routes } from 'react-router';
 
 import { Tabs } from '@polkadot/react-components';
 
@@ -36,14 +36,22 @@ function UtilitiesApp ({ basePath, className }: Props): React.ReactElement<Props
         basePath={basePath}
         items={tabsRef.current}
       />
-      <Switch>
-        <Route path={`${basePath}/hash`}>
-          <Hash />
+      <Routes>
+        <Route path={basePath}>
+          <Route
+            element={
+              <Hash />
+            }
+            path='hash'
+          />
+          <Route
+            element={
+              <Convert />
+            }
+            index
+          />
         </Route>
-        <Route>
-          <Convert />
-        </Route>
-      </Switch>
+      </Routes>
     </main>
   );
 }
