@@ -258,7 +258,7 @@ export const prodParasPolkadot: EndpointOption[] = [
     info: 'frequency',
     paraId: 2091,
     providers: {
-      'Frequency 0': 'wss://0.rpc.frequency.xyz',
+      // 'Frequency 0': 'wss://0.rpc.frequency.xyz', // https://github.com/polkadot-js/apps/issues/9231
       'Frequency 1': 'wss://1.rpc.frequency.xyz'
     },
     text: 'Frequency',
@@ -409,12 +409,9 @@ export const prodParasPolkadot: EndpointOption[] = [
   {
     homepage: 'https://manta.network',
     info: 'manta',
-    isUnreachable: true, // https://github.com/polkadot-js/apps/issues/7018
-    paraId: 2015,
+    paraId: 2104,
     providers: {
-      // 'Manta Kuhlii': 'wss://kuhlii.manta.systems', // https://github.com/polkadot-js/apps/issues/6930
-      // 'Manta Munkiana': 'wss://munkiana.manta.systems', // https://github.com/polkadot-js/apps/issues/6871
-      // 'Manta Pectinata': 'wss://pectinata.manta.systems' // https://github.com/polkadot-js/apps/issues/7018
+      'Manta Network': 'wss://ws.manta.systems'
     },
     text: 'Manta',
     ui: {
@@ -623,6 +620,7 @@ export const prodParasPolkadot: EndpointOption[] = [
     paraId: 2037,
     providers: {
       'Geo Load Balancer': 'wss://ws.unique.network',
+      OnFinality: 'wss://unique.api.onfinality.io/public-ws',
       'Unique America': 'wss://us-ws.unique.network',
       'Unique Asia': 'wss://asia-ws.unique.network',
       'Unique Europe': 'wss://eu-ws.unique.network'
