@@ -3,4 +3,4 @@
 
 // Do not edit, auto-generated by @polkadot/dev
 
-export const packageInfo = { name: '@polkadot/apps-config', path: 'auto', type: 'auto', version: '0.124.2-174-x' };
+export const packageInfo = { name: '@polkadot/apps-config', path: 'auto', type: 'auto', version: '0.124.2-192-x' };
