@@ -3,4 +3,4 @@
 
 // Do not edit. Auto-generated via node scripts/imgConvert.mjs
 
-export const chainsVaraSVG = 'data:image/svg+xml;base64,PD94bWwgdmVyc2lvbj0iMS4wIiBlbmNvZGluZz0idXRmLTgiPz4KPCEtLSBHZW5lcmF0b3I6IEFkb2JlIElsbHVzdHJhdG9yIDI2LjQuMSwgU1ZHIEV4cG9ydCBQbHVnLUluIC4gU1ZHIFZlcnNpb246IDYuMDAgQnVpbGQgMCkgIC0tPgo8c3ZnIHZlcnNpb249IjEuMSIgaWQ9IkxheWVyXzEiIHhtbG5zPSJodHRwOi8vd3d3LnczLm9yZy8yMDAwL3N2ZyIgeG1sbnM6eGxpbms9Imh0dHA6Ly93d3cudzMub3JnLzE5OTkveGxpbmsiIHg9IjBweCIgeT0iMHB4IgoJIHZpZXdCb3g9IjAgMCAyMDAgMjAwIiBzdHlsZT0iZW5hYmxlLWJhY2tncm91bmQ6bmV3IDAgMCAyMDAgMjAwOyIgeG1sOnNwYWNlPSJwcmVzZXJ2ZSI+CjxzdHlsZSB0eXBlPSJ0ZXh0L2NzcyI+Cgkuc3Qwe2ZpbGw6IzAwRkZCRjt9Cjwvc3R5bGU+CjxyZWN0IHdpZHRoPSIyMDAiIGhlaWdodD0iMjAwIi8+CjxnPgoJPHBhdGggY2xhc3M9InN0MCIgZD0iTTEzMy4yNywxMTEuN2wtMTYuMjEtNjEuMzJoLTMuNDJoLTMuNDJoLTMuNDJ2OTkuMjRoNi44NHYtMTguNGw2LjI4LTYuMjhsNi41MywyNC42OGw2LjgyLTAuMDVsLTcuOTQtMzAuMDQKCQlsMi41MS0yLjUxTDEzMy4yNywxMTEuN3ogTTExMy42NCwxMjEuNTVWNjMuMzFsMTIuMTgsNDYuMDZMMTEzLjY0LDEyMS41NXoiLz4KCTxwYXRoIGNsYXNzPSJzdDAiIGQ9Ik03OC42Niw1MC4zOGgtMy40MmgtMy40Mkg2OC40djk5LjI0aDYuODRsMTguMDQtMTguMDVsNC43NywxOC4wNWg2Ljg0TDc4LjY2LDUwLjM4eiBNNzUuMjQsMTM5Ljk0di05LjY4CgkJdi05LjY3di0xNC4wMVY2My4zMWwxMS45OCw0NS4zbDAsMGwyLjAyLDcuNjVsMCwwbDIuMDIsNy42Nkw3NS4yNCwxMzkuOTR6Ii8+Cgk8cGF0aCBjbGFzcz0ic3QwIiBkPSJNMTQ5LjQyLDUwLjM4aC0zLjQyaC0zLjQyaC0zLjQydjk5LjI0aDYuODRsMTguMDQtMTguMDVsNC43NywxOC4wNWg2Ljg0TDE0OS40Miw1MC4zOHogTTE0Ni4wMSwxMzkuOTR2LTkuNjgKCQl2LTkuNjd2LTEwLjk2VjYzLjMxbDEwLjYyLDQwLjE4bDAsMGwxLjIzLDQuNjZsMC4xMiwwLjQ2bDAsMGwyLjAyLDcuNjVsMCwwbDIuMDIsNy42NkwxNDYuMDEsMTM5Ljk0eiIvPgoJPHBvbHlnb24gY2xhc3M9InN0MCIgcG9pbnRzPSI1My45OSw1MC4zOCA1My45OSw4Ni45OSA1My45OSwxMTIuODUgNTMuOTksMTE4LjU4IDUzLjk5LDEzNi42OSAzMS4xOCw1MC4zOCAyNC4zNCw1MC4zOCAKCQk1MC41OCwxNDkuNjIgNTMuOTksMTQ5LjYyIDU3LjQxLDE0OS42MiA2MC44MywxNDkuNjIgNjAuODMsNTAuMzggCSIvPgo8L2c+Cjwvc3ZnPgo=';
+export const chainsVaraSVG = 'data:image/svg+xml;base64,PD94bWwgdmVyc2lvbj0iMS4wIiBlbmNvZGluZz0idXRmLTgiPz4KPCEtLSBHZW5lcmF0b3I6IEFkb2JlIElsbHVzdHJhdG9yIDI2LjQuMSwgU1ZHIEV4cG9ydCBQbHVnLUluIC4gU1ZHIFZlcnNpb246IDYuMDAgQnVpbGQgMCkgIC0tPgo8c3ZnIHZlcnNpb249IjEuMSIgaWQ9IkxheWVyXzEiIHhtbG5zPSJodHRwOi8vd3d3LnczLm9yZy8yMDAwL3N2ZyIgeG1sbnM6eGxpbms9Imh0dHA6Ly93d3cudzMub3JnLzE5OTkveGxpbmsiIHg9IjBweCIgeT0iMHB4IgoJIHZpZXdCb3g9IjAgMCA4MDAgODAwIiBzdHlsZT0iZW5hYmxlLWJhY2tncm91bmQ6bmV3IDAgMCA4MDAgODAwOyIgeG1sOnNwYWNlPSJwcmVzZXJ2ZSI+CjxzdHlsZSB0eXBlPSJ0ZXh0L2NzcyI+Cgkuc3Qwe2NsaXAtcGF0aDp1cmwoI1NWR0lEXzAwMDAwMTM3MTMxMDIzMTc1ODUxOTA5NjQwMDAwMDE1ODQzNDczNTM5MzY0Nzk0MjgzXyk7ZmlsbDojMDNGRkM0O30KCS5zdDF7Y2xpcC1wYXRoOnVybCgjU1ZHSURfMDAwMDAxMzcxMzEwMjMxNzU4NTE5MDk2NDAwMDAwMTU4NDM0NzM1MzkzNjQ3OTQyODNfKTt9Cjwvc3R5bGU+CjxnPgoJPGRlZnM+CgkJPHJlY3QgaWQ9IlNWR0lEXzFfIiB4PSIwIiB3aWR0aD0iODAwIiBoZWlnaHQ9IjgwMCIvPgoJPC9kZWZzPgoJPGNsaXBQYXRoIGlkPSJTVkdJRF8wMDAwMDAwODg2MTMzMDYxOTU1MDg5Mzg2MDAwMDAxMjMyMzcyNDM3NDAyMzQyNTY5NF8iPgoJCTx1c2UgeGxpbms6aHJlZj0iI1NWR0lEXzFfIiAgc3R5bGU9Im92ZXJmbG93OnZpc2libGU7Ii8+Cgk8L2NsaXBQYXRoPgoJCgkJPHJlY3Qgc3R5bGU9ImNsaXAtcGF0aDp1cmwoI1NWR0lEXzAwMDAwMDA4ODYxMzMwNjE5NTUwODkzODYwMDAwMDEyMzIzNzI0Mzc0MDIzNDI1Njk0Xyk7ZmlsbDojMDNGRkM0OyIgd2lkdGg9IjgwMCIgaGVpZ2h0PSI4MDAiLz4KCTxnIHN0eWxlPSJjbGlwLXBhdGg6dXJsKCNTVkdJRF8wMDAwMDAwODg2MTMzMDYxOTU1MDg5Mzg2MDAwMDAxMjMyMzcyNDM3NDAyMzQyNTY5NF8pOyI+CgkJPHJlY3QgeD0iNTM5LjkyIiB5PSItMTA2LjEyIiB3aWR0aD0iMTE2LjQyIiBoZWlnaHQ9IjEwMDQuOTciLz4KCTwvZz4KCTxnIHN0eWxlPSJjbGlwLXBhdGg6dXJsKCNTVkdJRF8wMDAwMDAwODg2MTMzMDYxOTU1MDg5Mzg2MDAwMDAxMjMyMzcyNDM3NDAyMzQyNTY5NF8pOyI+CgkJCgkJCTxyZWN0IHg9IjM3My45OSIgeT0iLTExMi43NSIgdHJhbnNmb3JtPSJtYXRyaXgoMC45MzYyIC0wLjM1MTQgMC4zNTE0IDAuOTM2MiAtOTguNzMyMyAxNzQuODIwOCkiIHdpZHRoPSIxMTYuNDIiIGhlaWdodD0iOTQ0LjI3Ii8+Cgk8L2c+Cgk8ZyBzdHlsZT0iY2xpcC1wYXRoOnVybCgjU1ZHSURfMDAwMDAwMDg4NjEzMzA2MTk1NTA4OTM4NjAwMDAwMTIzMjM3MjQzNzQwMjM0MjU2OTRfKTsiPgoJCQoJCQk8cmVjdCB4PSIxNDguODYiIHk9Ii0xMTYuMzQiIHRyYW5zZm9ybT0ibWF0cml4KDAuOTM2MiAtMC4zNTE0IDAuMzUxNCAwLjkzNjIgLTEzMS41NzggOTkuMDU0NSkiIHdpZHRoPSIxMTYuNDIiIGhlaWdodD0iMTA1Ni42MyIvPgoJPC9nPgo8L2c+Cjwvc3ZnPgo=';
