@@ -3,4 +3,4 @@
 
 // Do not edit. Auto-generated via node scripts/imgConvert.mjs
 
-export const nodesVaraSVG = 'data:image/svg+xml;base64,PD94bWwgdmVyc2lvbj0iMS4wIiBlbmNvZGluZz0idXRmLTgiPz4KPCEtLSBHZW5lcmF0b3I6IEFkb2JlIElsbHVzdHJhdG9yIDI2LjQuMSwgU1ZHIEV4cG9ydCBQbHVnLUluIC4gU1ZHIFZlcnNpb246IDYuMDAgQnVpbGQgMCkgIC0tPgo8c3ZnIHZlcnNpb249IjEuMSIgaWQ9IkxheWVyXzEiIHhtbG5zPSJodHRwOi8vd3d3LnczLm9yZy8yMDAwL3N2ZyIgeG1sbnM6eGxpbms9Imh0dHA6Ly93d3cudzMub3JnLzE5OTkveGxpbmsiIHg9IjBweCIgeT0iMHB4IgoJIHZpZXdCb3g9IjAgMCA4MDAgODAwIiBzdHlsZT0iZW5hYmxlLWJhY2tncm91bmQ6bmV3IDAgMCA4MDAgODAwOyIgeG1sOnNwYWNlPSJwcmVzZXJ2ZSI+CjxzdHlsZSB0eXBlPSJ0ZXh0L2NzcyI+Cgkuc3Qwe2NsaXAtcGF0aDp1cmwoI1NWR0lEXzAwMDAwMTM3MTMxMDIzMTc1ODUxOTA5NjQwMDAwMDE1ODQzNDczNTM5MzY0Nzk0MjgzXyk7ZmlsbDojMDNGRkM0O30KCS5zdDF7Y2xpcC1wYXRoOnVybCgjU1ZHSURfMDAwMDAxMzcxMzEwMjMxNzU4NTE5MDk2NDAwMDAwMTU4NDM0NzM1MzkzNjQ3OTQyODNfKTt9Cjwvc3R5bGU+CjxnPgoJPGRlZnM+CgkJPHJlY3QgaWQ9IlNWR0lEXzFfIiB4PSIwIiB3aWR0aD0iODAwIiBoZWlnaHQ9IjgwMCIvPgoJPC9kZWZzPgoJPGNsaXBQYXRoIGlkPSJTVkdJRF8wMDAwMDAwODg2MTMzMDYxOTU1MDg5Mzg2MDAwMDAxMjMyMzcyNDM3NDAyMzQyNTY5NF8iPgoJCTx1c2UgeGxpbms6aHJlZj0iI1NWR0lEXzFfIiAgc3R5bGU9Im92ZXJmbG93OnZpc2libGU7Ii8+Cgk8L2NsaXBQYXRoPgoJCgkJPHJlY3Qgc3R5bGU9ImNsaXAtcGF0aDp1cmwoI1NWR0lEXzAwMDAwMDA4ODYxMzMwNjE5NTUwODkzODYwMDAwMDEyMzIzNzI0Mzc0MDIzNDI1Njk0Xyk7ZmlsbDojMDNGRkM0OyIgd2lkdGg9IjgwMCIgaGVpZ2h0PSI4MDAiLz4KCTxnIHN0eWxlPSJjbGlwLXBhdGg6dXJsKCNTVkdJRF8wMDAwMDAwODg2MTMzMDYxOTU1MDg5Mzg2MDAwMDAxMjMyMzcyNDM3NDAyMzQyNTY5NF8pOyI+CgkJPHJlY3QgeD0iNTM5LjkyIiB5PSItMTA2LjEyIiB3aWR0aD0iMTE2LjQyIiBoZWlnaHQ9IjEwMDQuOTciLz4KCTwvZz4KCTxnIHN0eWxlPSJjbGlwLXBhdGg6dXJsKCNTVkdJRF8wMDAwMDAwODg2MTMzMDYxOTU1MDg5Mzg2MDAwMDAxMjMyMzcyNDM3NDAyMzQyNTY5NF8pOyI+CgkJCgkJCTxyZWN0IHg9IjM3My45OSIgeT0iLTExMi43NSIgdHJhbnNmb3JtPSJtYXRyaXgoMC45MzYyIC0wLjM1MTQgMC4zNTE0IDAuOTM2MiAtOTguNzMyMyAxNzQuODIwOCkiIHdpZHRoPSIxMTYuNDIiIGhlaWdodD0iOTQ0LjI3Ii8+Cgk8L2c+Cgk8ZyBzdHlsZT0iY2xpcC1wYXRoOnVybCgjU1ZHSURfMDAwMDAwMDg4NjEzMzA2MTk1NTA4OTM4NjAwMDAwMTIzMjM3MjQzNzQwMjM0MjU2OTRfKTsiPgoJCQoJCQk8cmVjdCB4PSIxNDguODYiIHk9Ii0xMTYuMzQiIHRyYW5zZm9ybT0ibWF0cml4KDAuOTM2MiAtMC4zNTE0IDAuMzUxNCAwLjkzNjIgLTEzMS41NzggOTkuMDU0NSkiIHdpZHRoPSIxMTYuNDIiIGhlaWdodD0iMTA1Ni42MyIvPgoJPC9nPgo8L2c+Cjwvc3ZnPgo=';
+export const nodesVaraSVG = 'data:image/svg+xml;base64,PD94bWwgdmVyc2lvbj0iMS4wIiBlbmNvZGluZz0idXRmLTgiPz4KPCEtLSBHZW5lcmF0b3I6IEFkb2JlIElsbHVzdHJhdG9yIDI2LjQuMSwgU1ZHIEV4cG9ydCBQbHVnLUluIC4gU1ZHIFZlcnNpb246IDYuMDAgQnVpbGQgMCkgIC0tPgo8c3ZnIHZlcnNpb249IjEuMSIgaWQ9IkxheWVyXzEiIHhtbG5zPSJodHRwOi8vd3d3LnczLm9yZy8yMDAwL3N2ZyIgeG1sbnM6eGxpbms9Imh0dHA6Ly93d3cudzMub3JnLzE5OTkveGxpbmsiIHg9IjBweCIgeT0iMHB4IgoJIHZpZXdCb3g9IjAgMCA4MDAgODAwIiBzdHlsZT0iZW5hYmxlLWJhY2tncm91bmQ6bmV3IDAgMCA4MDAgODAwOyIgeG1sOnNwYWNlPSJwcmVzZXJ2ZSI+CjxzdHlsZSB0eXBlPSJ0ZXh0L2NzcyI+Cgkuc3Qwe2NsaXAtcGF0aDp1cmwoI1NWR0lEXzAwMDAwMDI2MTQxNjA0NDcyMTM2NTA4NDAwMDAwMDA4MTQxNjUwMTgyMjA4MjM0NDE0Xyk7ZmlsbDojMDBCMzg3O30KCS5zdDF7Y2xpcC1wYXRoOnVybCgjU1ZHSURfMDAwMDAxNzc0NTMwODM0NTc0NjcyOTY3NTAwMDAwMDEyMzk2NTkyNTMwMjQwOTI4NDVfKTt9Cgkuc3Qye2ZpbGw6I0ZGRkZGRjt9Cgkuc3Qze2NsaXAtcGF0aDp1cmwoI1NWR0lEXzAwMDAwMDQ0MTczMjE2ODQ1MDY3Mzg3MzAwMDAwMDAxODY3ODQyMjQwNDgwMDk4MjA4Xyk7fQoJLnN0NHtjbGlwLXBhdGg6dXJsKCNTVkdJRF8wMDAwMDEzMzUyNjczNDg0MjI1OTU4NTExMDAwMDAwMTQ3MjQ0MzIyMDc5Njg1OTU2MV8pO30KPC9zdHlsZT4KPGc+Cgk8Zz4KCQk8ZGVmcz4KCQkJPHJlY3QgaWQ9IlNWR0lEXzFfIiB3aWR0aD0iODAwIiBoZWlnaHQ9IjgwMCIvPgoJCTwvZGVmcz4KCQk8Y2xpcFBhdGggaWQ9IlNWR0lEXzAwMDAwMDE4OTYwMDU0NTM0MTAyMzM4MDQwMDAwMDA4NjA5NTE3ODAxMTgyNzkxMzMwXyI+CgkJCTx1c2UgeGxpbms6aHJlZj0iI1NWR0lEXzFfIiAgc3R5bGU9Im92ZXJmbG93OnZpc2libGU7Ii8+CgkJPC9jbGlwUGF0aD4KCQkKCQkJPHJlY3Qgc3R5bGU9ImNsaXAtcGF0aDp1cmwoI1NWR0lEXzAwMDAwMDE4OTYwMDU0NTM0MTAyMzM4MDQwMDAwMDA4NjA5NTE3ODAxMTgyNzkxMzMwXyk7ZmlsbDojMDBCMzg3OyIgd2lkdGg9IjgwMCIgaGVpZ2h0PSI4MDAiLz4KCTwvZz4KCTxnPgoJCTxkZWZzPgoJCQk8cmVjdCBpZD0iU1ZHSURfMDAwMDAwMTAyODg5NjE0MzU2ODc0MzU3MTAwMDAwMTgxMTQ2NzA1MDkxNDMwOTA4NTZfIiB3aWR0aD0iODAwIiBoZWlnaHQ9IjgwMCIvPgoJCTwvZGVmcz4KCQk8Y2xpcFBhdGggaWQ9IlNWR0lEXzAwMDAwMDY5Mzk4MzU0NTE5MzQwODEyNzEwMDAwMDA4NTk0MzEyNjEyNDc2MzAwMjIxXyI+CgkJCTx1c2UgeGxpbms6aHJlZj0iI1NWR0lEXzAwMDAwMDEwMjg4OTYxNDM1Njg3NDM1NzEwMDAwMDE4MTE0NjcwNTA5MTQzMDkwODU2XyIgIHN0eWxlPSJvdmVyZmxvdzp2aXNpYmxlOyIvPgoJCTwvY2xpcFBhdGg+CgkJPGcgc3R5bGU9ImNsaXAtcGF0aDp1cmwoI1NWR0lEXzAwMDAwMDY5Mzk4MzU0NTE5MzQwODEyNzEwMDAwMDA4NTk0MzEyNjEyNDc2MzAwMjIxXyk7Ij4KCQkJPHJlY3QgeD0iNTM5LjkiIHk9Ii0xMDYuMSIgY2xhc3M9InN0MiIgd2lkdGg9IjExNi40IiBoZWlnaHQ9IjEwMDUiLz4KCQk8L2c+Cgk8L2c+Cgk8Zz4KCQk8ZGVmcz4KCQkJPHJlY3QgaWQ9IlNWR0lEXzAwMDAwMTMwNjQ3MTQyMTMyMjc2NzUyMzAwMDAwMDEyNjIwODY2MjYwMjA4MTA0MTE3XyIgd2lkdGg9IjgwMCIgaGVpZ2h0PSI4MDAiLz4KCQk8L2RlZnM+CgkJPGNsaXBQYXRoIGlkPSJTVkdJRF8wMDAwMDA0NTU5MzQ0MzU0NDM0MjgzNDY4MDAwMDAwMDI0ODMyMTAxMTgwNDY2MzIxNF8iPgoJCQk8dXNlIHhsaW5rOmhyZWY9IiNTVkdJRF8wMDAwMDEzMDY0NzE0MjEzMjI3Njc1MjMwMDAwMDAxMjYyMDg2NjI2MDIwODEwNDExN18iICBzdHlsZT0ib3ZlcmZsb3c6dmlzaWJsZTsiLz4KCQk8L2NsaXBQYXRoPgoJCTxnIHN0eWxlPSJjbGlwLXBhdGg6dXJsKCNTVkdJRF8wMDAwMDA0NTU5MzQ0MzU0NDM0MjgzNDY4MDAwMDAwMDI0ODMyMTAxMTgwNDY2MzIxNF8pOyI+CgkJCQoJCQkJPHJlY3QgeD0iMzc0IiB5PSItMTEyLjciIHRyYW5zZm9ybT0ibWF0cml4KDAuOTM2MiAtMC4zNTE0IDAuMzUxNCAwLjkzNjIgLTk4LjczMzMgMTc0Ljc5MzkpIiBjbGFzcz0ic3QyIiB3aWR0aD0iMTE2LjQiIGhlaWdodD0iOTQ0LjIiLz4KCQk8L2c+Cgk8L2c+Cgk8Zz4KCQk8ZGVmcz4KCQkJPHJlY3QgaWQ9IlNWR0lEXzAwMDAwMDA3NDAyOTAzMTIwODM5MDE1NjgwMDAwMDE4Mjg1MTM4MDIzOTM0NzgzOTA5XyIgd2lkdGg9IjgwMCIgaGVpZ2h0PSI4MDAiLz4KCQk8L2RlZnM+CgkJPGNsaXBQYXRoIGlkPSJTVkdJRF8wMDAwMDAwMzA2ODg2NTEwNDM4NTMyNzEwMDAwMDAwNTIyNTUzMzc3MDMwODEyNTYyNF8iPgoJCQk8dXNlIHhsaW5rOmhyZWY9IiNTVkdJRF8wMDAwMDAwNzQwMjkwMzEyMDgzOTAxNTY4MDAwMDAxODI4NTEzODAyMzkzNDc4MzkwOV8iICBzdHlsZT0ib3ZlcmZsb3c6dmlzaWJsZTsiLz4KCQk8L2NsaXBQYXRoPgoJCTxnIHN0eWxlPSJjbGlwLXBhdGg6dXJsKCNTVkdJRF8wMDAwMDAwMzA2ODg2NTEwNDM4NTMyNzEwMDAwMDAwNTIyNTUzMzc3MDMwODEyNTYyNF8pOyI+CgkJCQoJCQkJPHJlY3QgeD0iMTQ4LjgiIHk9Ii0xMTYuMyIgdHJhbnNmb3JtPSJtYXRyaXgoMC45MzYyIC0wLjM1MTQgMC4zNTE0IDAuOTM2MiAtMTMxLjU2ODUgOTkuMDMzOSkiIGNsYXNzPSJzdDIiIHdpZHRoPSIxMTYuNCIgaGVpZ2h0PSIxMDU2LjYiLz4KCQk8L2c+Cgk8L2c+CjwvZz4KPC9zdmc+Cg==';
