@@ -4,7 +4,7 @@
 import type { Hash } from '@polkadot/types/interfaces';
 import type { CodeJson, CodeStored } from './types.js';
 
-import EventEmitter from 'eventemitter3';
+import { EventEmitter } from 'eventemitter3';
 import store from 'store';
 
 import { Abi } from '@polkadot/api-contract';
