@@ -19,10 +19,10 @@ const BOB = '5FHneW46xGXgs5mUiveU4sbTyGBzmstUspZC92UhjJM694ty';
 const randomAmount = Math.floor((Math.random() * 100000) + 1);
 
 // Create a extrinsic, transferring randomAmount units to Bob.
-const transfer = api.tx.balances.transfer(BOB, randomAmount);
+const transfer = api.tx.balances.transferAllowDeath(BOB, randomAmount);
 
 // Sign and Send the transaction
-await transfer.signAndSend(ALICE, ({ events = [], status }) => {
+await transferAllowDeath.signAndSend(ALICE, ({ events = [], status }) => {
   if (status.isInBlock) {
     console.log('Successful transfer of ' + randomAmount + ' with hash ' + status.asInBlock.toHex());
   } else {
