@@ -10,7 +10,7 @@ const nonce = await api.query.system.accountNonce(ALICE),
 console.log('Current nonce', nonce);
 
 const unsub = await api.tx.balances
-  .transfer(recipient,12345)
+  .transferAllowDeath(recipient,12345)
   .signAndSend(sender, ({ events = [], status }) => {
     console.log('Transaction status:', status.type);
 
