@@ -4,7 +4,7 @@
 import type { ApiPromise } from '@polkadot/api';
 import type { SignedBlockExtended } from '@polkadot/api-derive/types';
 import type { Event } from '@polkadot/types/interfaces';
-import type { PolkadotPrimitivesV2CandidateReceipt } from '@polkadot/types/lookup';
+import type { PolkadotPrimitivesV4CandidateReceipt } from '@polkadot/types/lookup';
 import type { IEvent } from '@polkadot/types/types';
 import type { BN } from '@polkadot/util';
 import type { EventMapInfo } from './types.js';
@@ -26,7 +26,7 @@ const EMPTY_EVENTS: Result = { lastBacked: {}, lastIncluded: {}, lastTimeout: {}
 
 function includeEntry (map: EventMap, event: Event, blockHash: string, blockNumber: BN): void {
   try {
-    const { descriptor } = (event as unknown as IEvent<[PolkadotPrimitivesV2CandidateReceipt]>).data[0];
+    const { descriptor } = (event as unknown as IEvent<[PolkadotPrimitivesV4CandidateReceipt]>).data[0];
 
     if (descriptor && descriptor.paraId) {
       map[descriptor.paraId.toString()] = {
